Do not trim a mow path if it was not active
Occasionally mowing can be aborted (e.g low voltage or high temp) just before
a path is received by FTC planner, but FTC returns progress for the previous
path.
This progress is trimmed from the not-yet-started path, potentially skipping
the entire path.

This patch skips the trimming logic if the path is stil in pending/recalled
state.