Add function to create a default tasklist
mower_map should really support retrieving the number of areas.
Ideally, in the future there would be area attributes, e.g. whether to
include the area in the default tasklist and in which order, and there
would be a service returning the indexes of those areas in the correct
order.