Allow aborting Docking/Undocking
Pressing "Home" when undocking aborts it, and pressing "play" when docking aborts it.

Uninterruptable docking can be a real problem if your mower is stuck trying to return to the dock and you want to tele-operate it.  This adds button controls, as well as actions which can be used from the GUI.