Switch move_base_flex to post-Hydro params format to improve reliability
I was seeing flakiness in move_base_flex that seemed to be due to the automatic migration of parameters from the pre-Hydro to post-Hydro format (see http://wiki.ros.org/costmap_2d#costmap_2d.2Flayered.Pre-Hydro_Parameters).  It seems that sometimes the params would end up as a mixture of the two, causing the costmap to fail to be populated.

This PR just ports everything to the new post-Hydro format, so there is no automatic conversion that can go wrong.