Improve OM_RAIN_DELAY_MINUTES documentation (#183)
Just a minor improvement, but it took me some time to figure out and
maybe saves others some time (although a lower value is not reasonable
for normal operation I tried to lower it for testing and was surprised
that the value did not change).

Related info:
If the mower is in a garage the sensor dries although it is still
raining. After the delay period the mower goes out. If it is still
raining it goes into idle state at the end of undocking and not into
docking again. It stays there for ever. This problem is fixed by PR179
(I think the solution is to always go into mowing state and handle rain
detection there as always)