Expose rain_detected flag to robot_state/json MQTT topic (#210)
I am using Home Assistant to control when the mower goes out to mow or
when it has to return. I now managed to get the stock CoverUI working
with the rain sensor, but discovered that the sensor information is not
available through mqtt.

This PR makes the rain_detected flag available over MQTT, so home
assistant automations could use it.