Pete_MI632's lateral_priority_distance FTC tweak
This change adds a new parameter to the FTC planner, which allows *non-linear* control of the angle of the mower in responds to lateral errors.  This allows the angular PID to dominate when the lateral error is small, and the lateral PID to dominate when the lateral error is large.  The distance at which this switchover occurs is configured as `lateral_priority_distance`.  The angular PID will ramp down linearly from 100% gain to 10% gain as the cross track error increase from 0 to `lateral_priority_distance`.  Beyond `lateral_priority_distance` the gain will remain at 10%.