Publish capabilities via MQTT
The idea is to have a map of capabilities that make it easy for apps to
check what the mower supports, without having to compare versions.

For simplicity in usage, the capabilities should be represented by a
simple, flat object with the respective capabilities as keys and the
"generation" as value. It starts with 1 when the capability gets added
and increases whenever a significant change was done.